Added xmgrace output for pdos calculation similar to dos calculations
This labelling may not be entirely correct when using hand-selected projectors.
This is because all the labels for the projector (symbol, species, species_num and angular momentum)
are written out to the label even when some are superfluous.

Book-keeping is a bit of a pain for it so probablt easier to adjust plot by hand...